Split out addfeature as test util lib.
